+++ cosmos/client.go
@@ -19,14 +19,13 @@ import (
 	"github.com/mesg-foundation/engine/ext/xstrings"
 	"github.com/mesg-foundation/engine/hash"
 	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/node"
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	tenderminttypes "github.com/tendermint/tendermint/types"
 )
 
 // Client is a tendermint client with helper functions.
 type Client struct {
-	*rpcclient.Local
+	rpcclient.Client
 	cdc          *codec.Codec
 	kb           keys.Keybase
 	chainID      string
@@ -41,13 +40,13 @@ type Client struct {
 }
 
 // NewClient returns a rpc tendermint client.
-func NewClient(node *node.Node, cdc *codec.Codec, kb keys.Keybase, chainID, accName, accPassword, minGasPrices string) (*Client, error) {
+func NewClient(client rpcclient.Client, cdc *codec.Codec, kb keys.Keybase, chainID, accName, accPassword, minGasPrices string) (*Client, error) {
 	minGasPricesDecoded, err := sdktypes.ParseDecCoins(minGasPrices)
 	if err != nil {
 		return nil, err
 	}
 	return &Client{
-		Local:        rpcclient.NewLocal(node),
+		Client:       client,
 		cdc:          cdc,
 		kb:           kb,
 		chainID:      chainID,
@@ -93,7 +92,7 @@ func (c *Client) QueryWithData(path string, data []byte) ([]byte, int64, error) 
 // BuildAndBroadcastMsg builds and signs message and broadcast it to node.
 func (c *Client) BuildAndBroadcastMsg(msg sdktypes.Msg) (*abci.ResponseDeliverTx, error) {
 	c.broadcastMutex.Lock() // Lock the whole signature + broadcast of the transaction
-	signedTx, err := c.createAndSignTx([]sdktypes.Msg{msg})
+	signedTx, err := c.CreateAndSignTx([]sdktypes.Msg{msg})
 	if err != nil {
 		c.broadcastMutex.Unlock()
 		return nil, err
@@ -106,7 +105,7 @@ func (c *Client) BuildAndBroadcastMsg(msg sdktypes.Msg) (*abci.ResponseDeliverTx
 	}
 
 	if txres.Code != abci.CodeTypeOK {
-		return nil, fmt.Errorf("transaction returned with invalid code %d", txres.Code)
+		return nil, fmt.Errorf("transaction returned with invalid code %d: %s", txres.Code, txres.Log)
 	}
 
 	// TODO: 20*time.Second should not be hardcoded here
@@ -132,7 +131,7 @@ func (c *Client) BuildAndBroadcastMsg(msg sdktypes.Msg) (*abci.ResponseDeliverTx
 		}
 		return &data.TxResult.Result, nil
 	case <-ctx.Done():
-		return nil, errors.New("i/o timeout")
+		return nil, errors.New("reach timeout for listening for transaction result")
 	}
 }
 
@@ -204,7 +203,8 @@ func (c *Client) GetAccount() (authExported.Account, error) {
 	return c.acc, nil
 }
 
-func (c *Client) createAndSignTx(msgs []sdktypes.Msg) (tenderminttypes.Tx, error) {
+// CreateAndSignTx build and sign a msg with client account.
+func (c *Client) CreateAndSignTx(msgs []sdktypes.Msg) (tenderminttypes.Tx, error) {
 	// retrieve account
 	accR, err := c.GetAccount()
 	if err != nil {