+++ e2e/runner_test.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/bank"
 	"github.com/mesg-foundation/engine/hash"
 	"github.com/mesg-foundation/engine/protobuf/acknowledgement"
 	pb "github.com/mesg-foundation/engine/protobuf/api"
@@ -12,6 +14,7 @@ import (
 )
 
 var testRunnerHash hash.Hash
+var testRunnerAddress sdk.AccAddress
 
 func testRunner(t *testing.T) {
 	t.Run("create", func(t *testing.T) {
@@ -35,6 +38,17 @@ func testRunner(t *testing.T) {
 		require.NoError(t, err)
 	})
 
+	t.Run("recreate", func(t *testing.T) {
+		_, err := client.RunnerClient.Delete(context.Background(), &pb.DeleteRunnerRequest{Hash: testRunnerHash})
+		require.NoError(t, err)
+		resp, err := client.RunnerClient.Create(context.Background(), &pb.CreateRunnerRequest{
+			ServiceHash: testServiceHash,
+			Env:         []string{"BAR=3", "REQUIRED=4"},
+		})
+		require.NoError(t, err)
+		testRunnerHash = resp.Hash
+	})
+
 	t.Run("get", func(t *testing.T) {
 		t.Run("grpc", func(t *testing.T) {
 			resp, err := client.RunnerClient.Get(context.Background(), &pb.GetRunnerRequest{Hash: testRunnerHash})
@@ -46,6 +60,7 @@ func testRunner(t *testing.T) {
 			var r *runner.Runner
 			lcdGet(t, "runner/get/"+testRunnerHash.String(), &r)
 			require.Equal(t, testRunnerHash, r.Hash)
+			testRunnerAddress = r.Address
 		})
 	})
 
@@ -81,5 +96,11 @@ func testDeleteRunner(t *testing.T) {
 		rs := make([]*runner.Runner, 0)
 		lcdGet(t, "runner/list", &rs)
 		require.Len(t, rs, 0)
+	})
+	t.Run("check coins on runner", func(t *testing.T) {
+		var coins sdk.Coins
+		param := bank.NewQueryBalanceParams(testRunnerAddress)
+		require.NoError(t, cclient.QueryJSON("custom/bank/balances", param, &coins))
+		require.True(t, coins.IsZero(), coins)
 	})
 }