+++ x/execution/client/rest/query.go
@@ -37,7 +37,7 @@ func queryGetHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}
 
@@ -63,7 +63,7 @@ func queryListHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}