+++ x/ownership/internal/keeper/keeper.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/mesg-foundation/engine/hash"
 	"github.com/mesg-foundation/engine/ownership"
 	"github.com/mesg-foundation/engine/x/ownership/internal/types"
@@ -15,13 +16,16 @@ import (
 type Keeper struct {
 	storeKey sdk.StoreKey
 	cdc      *codec.Codec
+
+	bankKeeper types.BankKeeper
 }
 
 // NewKeeper creates a ownership keeper
-func NewKeeper(cdc *codec.Codec, key sdk.StoreKey) Keeper {
+func NewKeeper(cdc *codec.Codec, key sdk.StoreKey, bankKeeper types.BankKeeper) Keeper {
 	keeper := Keeper{
-		storeKey: key,
-		cdc:      cdc,
+		storeKey:   key,
+		cdc:        cdc,
+		bankKeeper: bankKeeper,
 	}
 	return keeper
 }
@@ -42,7 +46,7 @@ func (k *Keeper) List(ctx sdk.Context) ([]*ownership.Ownership, error) {
 	for iter.Valid() {
 		var o *ownership.Ownership
 		if err := k.cdc.UnmarshalBinaryLengthPrefixed(iter.Value(), &o); err != nil {
-			return nil, err
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrJSONUnmarshal, err.Error())
 		}
 		ownerships = append(ownerships, o)
 		iter.Next()
@@ -51,54 +55,90 @@ func (k *Keeper) List(ctx sdk.Context) ([]*ownership.Ownership, error) {
 }
 
 // Set creates a new ownership.
-func (k Keeper) Set(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash, resource ownership.Ownership_Resource) (*ownership.Ownership, error) {
+func (k Keeper) Set(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash, resource ownership.Ownership_Resource, resourceAddress sdk.AccAddress) (*ownership.Ownership, error) {
 	store := ctx.KVStore(k.storeKey)
-	hashes := k.findOwnerships(store, "", resourceHash)
-	if len(hashes) > 0 {
-		return nil, fmt.Errorf("resource %s:%q already has an owner", resource, resourceHash)
+	own, err := k.getOwnership(store, resourceHash)
+	if err != nil {
+		return nil, err
+	}
+	if own != nil {
+		return nil, sdkerrors.Wrapf(sdkerrors.ErrUnauthorized, "resource %s:%q already has an owner", resource, resourceHash)
 	}
 
-	ownership := &ownership.Ownership{
-		Owner:        owner.String(),
-		Resource:     resource,
-		ResourceHash: resourceHash,
+	// TODO: should be moved to a new function
+	own = &ownership.Ownership{
+		Owner:           owner.String(),
+		Resource:        resource,
+		ResourceHash:    resourceHash,
+		ResourceAddress: resourceAddress,
 	}
-	ownership.Hash = hash.Dump(ownership)
+	own.Hash = hash.Dump(own)
 
-	store.Set(ownership.Hash, k.cdc.MustMarshalBinaryLengthPrefixed(ownership))
-	return ownership, nil
+	data, err := k.cdc.MarshalBinaryLengthPrefixed(own)
+	if err != nil {
+		return nil, sdkerrors.Wrapf(sdkerrors.ErrJSONMarshal, err.Error())
+	}
+	store.Set(own.Hash, data)
+	return own, nil
 }
 
 // Delete deletes an ownership.
 func (k Keeper) Delete(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash) error {
 	store := ctx.KVStore(k.storeKey)
-	hashes := k.findOwnerships(store, owner.String(), resourceHash)
-	if len(hashes) == 0 {
-		return fmt.Errorf("resource %q do not have any ownership", resourceHash)
+	own, err := k.getOwnership(store, resourceHash)
+	if err != nil {
+		return err
+	}
+	if own == nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "resource %q do not have any ownership", resourceHash)
+	}
+	if own.Owner != owner.String() {
+		return sdkerrors.Wrapf(sdkerrors.ErrUnauthorized, "resource %q is not owned by you", resourceHash)
 	}
 
-	for _, hash := range hashes {
-		store.Delete(hash)
+	// transfer all spendable coins from resource address to owner
+	coins := k.bankKeeper.GetCoins(ctx, own.ResourceAddress)
+	if !coins.IsZero() {
+		if err := k.bankKeeper.SendCoins(ctx, own.ResourceAddress, owner, coins); err != nil {
+			return err
+		}
 	}
+
+	// remove ownership
+	store.Delete(own.Hash)
 	return nil
 }
 
-// hasOwner checks if given resource has owner. Returns all ownership hash and true if has one
-// nil and false otherwise.
-func (k Keeper) findOwnerships(store sdk.KVStore, owner string, resourceHash hash.Hash) []hash.Hash {
-	var ownerships []hash.Hash
-	iter := store.Iterator(nil, nil)
+// WithdrawCoins try to withdraw coins to owner rom specific resource.
+func (k Keeper) WithdrawCoins(ctx sdk.Context, msg types.MsgWithdrawCoins) error {
+	store := ctx.KVStore(k.storeKey)
+	own, err := k.getOwnership(store, msg.Hash)
+	if err != nil {
+		return err
+	}
+	if own == nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "resource %q does not have any ownership", msg.Hash)
+	}
+	if own.Owner != msg.Owner.String() {
+		return sdkerrors.Wrapf(sdkerrors.ErrUnauthorized, "resource %q is not owned by you", msg.Hash)
+	}
+	return k.bankKeeper.SendCoins(ctx, own.ResourceAddress, msg.Owner, msg.Amount)
+}
 
+// getOwnership returns the ownership of a given resource.
+func (k Keeper) getOwnership(store sdk.KVStore, resourceHash hash.Hash) (*ownership.Ownership, error) {
+	iter := store.Iterator(nil, nil)
+	var own *ownership.Ownership
 	for iter.Valid() {
-		var o *ownership.Ownership
-		if err := k.cdc.UnmarshalBinaryLengthPrefixed(iter.Value(), &o); err == nil {
-			if (owner == "" || o.Owner == owner) && o.ResourceHash.Equal(resourceHash) {
-				ownerships = append(ownerships, o.Hash)
-			}
+		if err := k.cdc.UnmarshalBinaryLengthPrefixed(iter.Value(), &own); err != nil {
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrJSONUnmarshal, err.Error())
+		}
+		if own.ResourceHash.Equal(resourceHash) {
+			iter.Close()
+			return own, nil
 		}
 		iter.Next()
 	}
-
 	iter.Close()
-	return ownerships
+	return nil, nil
 }