+++ x/ownership/internal/types/expected_keepers.go
@@ -12,3 +12,9 @@ type ParamSubspace interface {
 	GetParamSet(ctx sdk.Context, ps params.ParamSet)
 	SetParamSet(ctx sdk.Context, ps params.ParamSet)
 }
+
+// BankKeeper module interface.
+type BankKeeper interface {
+	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) error
+	GetCoins(ctx sdk.Context, addr sdk.AccAddress) sdk.Coins
+}