+++ x/process/internal/types/expected_keepers.go
@@ -24,5 +24,10 @@ type InstanceKeeper interface {
 // OwnershipKeeper module interface.
 type OwnershipKeeper interface {
 	Delete(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash) error
-	Set(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash, resource ownershippb.Ownership_Resource) (*ownershippb.Ownership, error)
+	Set(ctx sdk.Context, owner sdk.AccAddress, resourceHash hash.Hash, resource ownershippb.Ownership_Resource, resourceAddress sdk.AccAddress) (*ownershippb.Ownership, error)
+}
+
+// BankKeeper module interface.
+type BankKeeper interface {
+	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) error
 }