+++ x/runner/handler.go
@@ -1,8 +1,6 @@
 package runner
 
 import (
-	"fmt"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/mesg-foundation/engine/x/runner/internal/types"
@@ -18,8 +16,7 @@ func NewHandler(k Keeper) sdk.Handler {
 		case MsgDeleteRunner:
 			return handleMsgDeleteRunner(ctx, k, &msg)
 		default:
-			errMsg := fmt.Sprintf("unrecognized %s message type: %T", ModuleName, msg)
-			return nil, sdkerrors.Wrap(sdkerrors.ErrUnknownRequest, errMsg)
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unrecognized %s message type: %T", types.ModuleName, msg)
 		}
 	}
 }