+++ x/runner/internal/types/msg.go
@@ -3,13 +3,13 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/mesg-foundation/engine/cosmos/errors"
 	"github.com/mesg-foundation/engine/ext/xvalidator"
 	"github.com/mesg-foundation/engine/hash"
 )
 
 // MsgCreateRunner defines a state transition to create a runner.
 type MsgCreateRunner struct {
+	// TODO: address should be renamed to owner
 	Address     sdk.AccAddress `json:"address" validate:"required,accaddress"`
 	ServiceHash hash.Hash      `json:"serviceHash" validate:"required,hash"`
 	EnvHash     hash.Hash      `json:"envHash" validate:"omitempty,hash"`
@@ -37,16 +37,16 @@ func (msg MsgCreateRunner) Type() string {
 // ValidateBasic runs stateless checks on the message.
 func (msg MsgCreateRunner) ValidateBasic() error {
 	if err := xvalidator.Validate.Struct(msg); err != nil {
-		return err
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, err.Error())
 	}
 	if msg.ServiceHash.IsZero() {
-		return sdkerrors.Wrap(errors.ErrValidation, "serviceHash is missing")
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "serviceHash is missing")
 	}
 	if msg.EnvHash.IsZero() {
-		return sdkerrors.Wrap(errors.ErrValidation, "envHash is missing")
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "envHash is missing")
 	}
 	if msg.Address.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "address is missing")
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "address is missing")
 	}
 	return nil
 }
@@ -88,13 +88,13 @@ func (msg MsgDeleteRunner) Type() string {
 // ValidateBasic runs stateless checks on the message.
 func (msg MsgDeleteRunner) ValidateBasic() error {
 	if err := xvalidator.Validate.Struct(msg); err != nil {
-		return sdkerrors.Wrap(errors.ErrValidation, err.Error())
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, err.Error())
 	}
 	if msg.RunnerHash.IsZero() {
-		return sdkerrors.Wrap(errors.ErrValidation, "runnerHash is missing")
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "runnerHash is missing")
 	}
 	if msg.Address.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "address is missing")
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "address is missing")
 	}
 	return nil
 }