+++ x/service/client/rest/query.go
@@ -44,7 +44,7 @@ func queryGetHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}
 
@@ -64,7 +64,7 @@ func queryListHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}
 
@@ -112,7 +112,7 @@ func queryExistHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}